// src/app/shared/models/team-leader.ts
export class TeamLeader {
  'avatar'?: string;
  'email'?: string;
  'first_name'?: string;
  'last_name'?: string;
  'username'?: string;
  'password'?: string;
  'position'?: string;
  'team'?: number;
  'department'?: number;
  'role'?: string;
  'gender'?: string;
  'birth_date'?: Date;
  'contact'?: string;
  'maritalStatus'?: string;
  'address'?: string;
  'joinDate'?: string;
  'enterprise'?: number;
  'social_media'?: {
    facebook: string;
    twitter: string;
    linkedin: string;
    discord: string;
  };
}